[MLIR] Fix circular dependency by moving StateStack to IR folder
This commit resolves a circular dependency issue between mlir/Support and mlir/IR:

- Move StateStack.h and StateStack.cpp from Support to IR folder
- Update CMakeLists.txt files to reflect the new locations
- Update Bazel BUILD file to maintain correct dependencies
- Update includes in affected files (flang, Target/LLVMIR)

The circular dependency was caused by StateStack.h depending on IR/Visitors.h
while other IR files depended on Support. Moving StateStack to IR eliminates
this cycle while maintaining proper separation of concerns.

This change improves build performance and reduces compilation issues.
