+++ utils/bazel/llvm-project-overlay/mlir/BUILD.bazel
@@ -4845,11 +4845,7 @@ cc_library(
     ]),
     hdrs = glob(["include/mlir/Support/*.h"]),
     includes = ["include"],
-    deps = [
-        "//llvm:Support",
-        "//mlir:IR",
-    ],
-)
+    deps = ["//llvm:Support"],)
 
 cc_library(
     name = "Debug",
